refactor: SurveyDomainService Command, Query로 분리
