+++ project/sangkyeong-onboarding/application/src/main/kotlin/org/survey/application/usecase/survey/CreateSurveyUseCase.kt
@@ -4,11 +4,11 @@ import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import org.survey.application.dto.command.CreateSurveyCommand
 import org.survey.domain.dto.SurveyItemData
-import org.survey.domain.survey.service.SurveyDomainService
+import org.survey.domain.survey.service.SurveyCommandService
 
 @Service
 class CreateSurveyUseCase(
-    private val surveyDomainService: SurveyDomainService,
+    private val surveyCommandService: SurveyCommandService,
 ) {
     @Transactional
     fun execute(command: CreateSurveyCommand) {
@@ -23,6 +23,6 @@ class CreateSurveyUseCase(
                 )
             }
 
-        surveyDomainService.createSurvey(command.title, command.description, surveyItemDataList)
+        surveyCommandService.createSurvey(command.title, command.description, surveyItemDataList)
     }
 }
